requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.4
distro==1.9.0
Django==4.2.10
exceptiongroup==1.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
markdown2==2.4.12
packaging==23.2
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pypdfium2==4.27.0
pytesseract==0.3.10
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
